requirements.txt
Flask==2.3.3
PyPDF2==3.0.1
python-dotenv==1.0.0
bcrypt==4.0.1
mysql-connector-python==8.1.0
youtube-transcript-api==0.6.1
beautifulsoup4==4.12.2
requests==2.31.0
reportlab==4.0.4
gunicorn==21.2.0